Update Makefile
Changes Makefile to properly recognize the current conda environment filepaths (and likewise nvcc). Furthermore, now compiles for additional architectures (sm_70, 75, 80, and 86 if user has at least CUDA 11, and does 89 and 90 if nvcc --version returns CUDA 12). For maintenance of compatibility with CUDA 11 and 12, adds both lib64 and lib in addition to compiling with C++ 17 instead of 20.